requirements.txt
python-jose~=3.3.0
passlib~=1.7.4
python-dotenv~=1.0.1
SQLAlchemy~=2.0.32
fastapi~=0.112.2
pydantic~=2.8.2
alembic~=1.13.2
schemas~=0.7.1